DEAR-132 convert dto responses to integers

+++ src/main/java/ch/fhnw/deardevbackend/dto/insights/WorkKindCountPerDayInsightDTO.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class WorkKindCountPerDayInsightDTO {
     private Integer workKindCount;
-    private Double userAverageHappiness;
-    private Double teamAverageHappiness;
+    private Integer userAverageHappiness;
+    private Integer teamAverageHappiness;
 }
